Remove redundant casts and centralize more structs
- Remove ~280 redundant casts from centralized constants
  (LZMA_VLI_UNKNOWN as lzma_vli, LZMA_FILTER_* as lzma_vli,
  UINT64_MAX as u64, LZMA_MEMUSAGE_BASE as u64)
- Remove ~170 redundant 'as lzma_ret' casts on functions
  that already return lzma_ret
- Move lzma_options_lzma, lzma_internal_s, lzma_stream structs
  and ISEQ_* constants into types.rs from 19 files each
- Rename C2RustUnnamed to lzma_internal_sequence in types.rs

+++ liblzma-rs/src/common/outqueue.rs
@@ -47,11 +47,11 @@ extern "C" fn lzma_outq_outbuf_memusage(buf_size: size_t) -> u64 {
 }
 #[no_mangle]
 pub extern "C" fn lzma_outq_memusage(buf_size_max: u64, threads: u32) -> u64 {
-    let limit: u64 = (UINT64_MAX as u64)
+    let limit: u64 = (UINT64_MAX)
         .wrapping_div((2 as c_int * 16384 as c_int) as u64)
         .wrapping_div(2);
     if threads > LZMA_THREADS_MAX as u32 || buf_size_max > limit {
-        return UINT64_MAX as u64;
+        return UINT64_MAX;
     }
     return ((2u32).wrapping_mul(threads) as u64)
         .wrapping_mul(lzma_outq_outbuf_memusage(buf_size_max as size_t));
